BACKEND/requirements/base.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0

# Database
psycopg2-binary>=2.9

# Async / Tasks
celery>=5.4
django-celery-beat>=2.6
flower>=2.0
redis>=5.0
channels>=4.0
channels-redis>=4.2
daphne>=4.0

# HTTP clients
requests>=2.31
cryptography>=42.0

# API Docs
drf-spectacular>=0.27

# AI (Phase 4)
openai>=1.40
numpy>=1.26
pypdf>=4.0
Pillow>=10.0

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.24
